feat(py13): upgrade Python 3.9 -> 3.13
- Remove labber git dependency (lib/zcu_tools/utils/labber_io.py replaces it;
  codebase already had zero direct Labber imports) and fix the bare import in
  datasaver.load_comment; anchor labber_io with 35 round-trip tests + 4
  datasaver wrapper tests; fix _read_trace_log stacked reshape for
  equal-length multi-entry traces with no outer axis.
- Drop dead server/design extras and the Labber-era numpy==1.23.5 /
  matplotlib==3.9 pins; resolve to numpy 2.4.6 / matplotlib 3.10.9 /
  pandas 3.0.3 / PyQt6 6.11 on Python 3.13 (.python-version 3.13,
  requires-python >=3.13). ZCU board stays on its own Python 3.8 PYNQ env
  (script/start_server.py unchanged).
- Fix typing fallout from numpy 2 / 3.13 stricter stubs (pyright 83+36 -> 0):
  honest float()/asarray conversions, runner child_type overloads, narrow
  ignores only for mpl/qutip stub gaps.
- Fix three stale GUI test fixtures whose in-slot exceptions aborted under
  PyQt (MagicMock vs FakeDeviceInfo/YOKOGS200Info; outdated
  get_persisted_left_panel_width mock) and add quiesce fixture to
  test_background.py (3.13 GC exposed queued-signal use-after-free).
- Verified: pytest 2594 passed, pyright 0 errors, ruff clean, Pyro4 smoke ok,
  three GUIs (measure/fluxdep/dispersive) launch + state_check over MCP.

+++ lib/zcu_tools/experiment/utils/single_shot/ge/base.py
@@ -127,7 +127,7 @@ def calc_fidelity(
     tn, fn = ng[:tind].sum(), ne[:tind].sum()
     fid = fidelity_func(tp, tn, fp, fn)
 
-    return fid, 0.5 * (bins[tind] + bins[tind + 1])
+    return fid, float(0.5 * (bins[tind] + bins[tind + 1]))
 
 
 def calc_ideal_fidelity(sg: float, se: float, s: float) -> float:
